+++ src/pages/128x64x1/main_page.c
@@ -96,7 +96,7 @@ void PAGE_MainExit()
 {
 }
 
-static u8 _action_cb(u32 button, u8 flags, void *data)
+static unsigned _action_cb(u32 button, unsigned flags, void *data)
 {
     u8 i;
     if ((flags & BUTTON_PRESS) && CHAN_ButtonIsPressed(button, BUT_ENTER)) {
