+++ src/pages/128x64x1/telemtest_page.c
@@ -27,7 +27,7 @@ typedef enum {
     telemetry_off,
 } TeleMetryMonitorType;
 
-static u8 _action_cb(u32 button, u8 flags, void *data);
+static unsigned _action_cb(u32 button, unsigned flags, void *data);
 static void _press_cb(guiObject_t *obj, const void *data);
 static const char *idx_cb(guiObject_t *obj, const void *data);
 
@@ -316,7 +316,7 @@ static void _navigate_pages(s8 direction)
     }
 }
 
-static u8 _action_cb(u32 button, u8 flags, void *data)
+static unsigned _action_cb(u32 button, unsigned flags, void *data)
 {
     (void)data;
     if ((flags & BUTTON_PRESS) || (flags & BUTTON_LONGPRESS)) {
